common component setup
