+++ v2/client/src/components/common/Feedback/Feedback.style.js
@@ -1,0 +1,38 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  max-width: 900px;
+  margin: 0 auto;
+  padding-top: 50px;
+`;
+
+export const Container = styled.div`
+  width: 98%;
+`;
+
+export const HeadlineDiv = styled.div`
+  margin-left: 20px;
+`;
+
+export const Paragraph = styled.p`
+  font-family: Roboto;
+  font-style: italic;
+  font-size: 30px;
+  line-height: 37px;
+  text-align: center;
+  margin-bottom: 30px;
+  color: #828282;
+`;
+export const ChartWrapper = styled.div`
+  margin-bottom: 50px;
+  padding: 20px 0;
+`;
+
+export const Description = styled.p`
+  font-family: Roboto;
+  font-size: 16px;
+  line-height: 19px;
+
+  color: #4f4f4f;
+  margin-bottom: 20px;
+`;
